add base recovery mechanism
Before this change, we stored chapters content in memory and we lost all
data in case of any crash. Now we store the content of the chapter in
the corresponded file, and it allows us to restore the job with the next
run or call.

Also, we use much less memory which could be a significant point in the
case of large books.